refactor(webmcp): simplify authoring progress and showcase
